# PERF: ascii c string functions