--- a/pandas/_libs/src/headers/portable.h
+++ b/pandas/_libs/src/headers/portable.h
@@ -5,4 +5,10 @@
 #define strcasecmp( s1, s2 ) _stricmp( s1, s2 )
 #endif
 
+// GH-23516 - works around locale perf issues
+// from MUSL libc, MIT Licensed - see LICENSES
+#define isdigit_ascii(c) ((unsigned)c - '0' < 10)
+#define isspace_ascii(c) (c == ' ' || (unsigned)c-'\t' < 5)
+#define toupper_ascii(c) (((unsigned)c-'a' < 26) ? (c & 0x5f) : c)
+
 #endif