--- a/pandas/_libs/src/parse_helper.h
+++ b/pandas/_libs/src/parse_helper.h
@@ -138,11 +138,11 @@ parsingerror:
 //
 
 PANDAS_INLINE void lowercase(char *p) {
-    for (; *p; ++p) *p = tolower(*p);
+    for (; *p; ++p) *p = tolower_ascii(*p);
 }
 
 PANDAS_INLINE void uppercase(char *p) {
-    for (; *p; ++p) *p = toupper(*p);
+    for (; *p; ++p) *p = toupper_ascii(*p);
 }
 
 static double xstrtod(const char *str, char **endptr, char decimal, char sci,
@@ -207,7 +207,7 @@ static double xstrtod(const char *str, char **endptr, char decimal, char sci,
     if (negative) number = -number;
 
     // Process an exponent string
-    if (toupper(*p) == toupper(sci)) {
+    if (toupper_ascii(*p) == toupper_ascii(sci)) {
         *maybe_int = 0;
 
         // Handle optional sign
@@ -263,7 +263,7 @@ static double xstrtod(const char *str, char **endptr, char decimal, char sci,
 
     if (skip_trailing) {
         // Skip trailing whitespace
-        while (isspace(*p)) p++;
+        while (isspace_ascii(*p)) p++;
     }
 
     if (endptr) *endptr = p;