--- a/pandas/_libs/src/parser/tokenizer.c
+++ b/pandas/_libs/src/parser/tokenizer.c
@@ -23,6 +23,8 @@ GitHub. See Python Software Foundation License and BSD licenses for these.
 #include <float.h>
 #include <math.h>
 
+#include "../headers/portable.h" // for isdigit_ascii
+
 static void *safe_realloc(void *buffer, size_t size) {
     void *result;
     // OSX is weird.
@@ -1411,7 +1413,7 @@ int tokenize_all_rows(parser_t *self) {
 }
 
 PANDAS_INLINE void uppercase(char *p) {
-    for (; *p; ++p) *p = toupper(*p);
+    for (; *p; ++p) *p = toupper_ascii(*p);
 }
 
 int PANDAS_INLINE to_longlong(char *item, long long *p_value) {
@@ -1424,7 +1426,7 @@ int PANDAS_INLINE to_longlong(char *item, long long *p_value) {
     *p_value = strtoll(item, &p_end, 10);
 
     // Allow trailing spaces.
-    while (isspace(*p_end)) ++p_end;
+    while (isspace_ascii(*p_end)) ++p_end;
 
     return (errno == 0) && (!*p_end);
 }
@@ -1541,7 +1543,7 @@ double xstrtod(const char *str, char **endptr, char decimal, char sci,
     errno = 0;
 
     // Skip leading whitespace.
-    while (isspace(*p)) p++;
+    while (isspace_ascii(*p)) p++;
 
     // Handle optional sign.
     negative = 0;
@@ -1558,7 +1560,7 @@ double xstrtod(const char *str, char **endptr, char decimal, char sci,
     num_decimals = 0;
 
     // Process string of digits.
-    while (isdigit(*p)) {
+    while (isdigit_ascii(*p)) {
         number = number * 10. + (*p - '0');
         p++;
         num_digits++;
@@ -1570,7 +1572,7 @@ double xstrtod(const char *str, char **endptr, char decimal, char sci,
     if (*p == decimal) {
         p++;
 
-        while (isdigit(*p)) {
+        while (isdigit_ascii(*p)) {
             number = number * 10. + (*p - '0');
             p++;
             num_digits++;
@@ -1589,7 +1591,7 @@ double xstrtod(const char *str, char **endptr, char decimal, char sci,
     if (negative) number = -number;
 
     // Process an exponent string.
-    if (toupper(*p) == toupper(sci)) {
+    if (toupper_ascii(*p) == toupper_ascii(sci)) {
         // Handle optional sign.
         negative = 0;
         switch (*++p) {
@@ -1602,7 +1604,7 @@ double xstrtod(const char *str, char **endptr, char decimal, char sci,
         // Process string of digits.
         num_digits = 0;
         n = 0;
-        while (isdigit(*p)) {
+        while (isdigit_ascii(*p)) {
             n = n * 10 + (*p - '0');
             num_digits++;
             p++;
@@ -1643,7 +1645,7 @@ double xstrtod(const char *str, char **endptr, char decimal, char sci,
 
     if (skip_trailing) {
         // Skip trailing whitespace.
-        while (isspace(*p)) p++;
+        while (isspace_ascii(*p)) p++;
     }
 
     if (endptr) *endptr = p;
@@ -1697,7 +1699,7 @@ double precise_xstrtod(const char *str, char **endptr, char decimal, char sci,
     errno = 0;
 
     // Skip leading whitespace.
-    while (isspace(*p)) p++;
+    while (isspace_ascii(*p)) p++;
 
     // Handle optional sign.
     negative = 0;
@@ -1714,7 +1716,7 @@ double precise_xstrtod(const char *str, char **endptr, char decimal, char sci,
     num_decimals = 0;
 
     // Process string of digits.
-    while (isdigit(*p)) {
+    while (isdigit_ascii(*p)) {
         if (num_digits < max_digits) {
             number = number * 10. + (*p - '0');
             num_digits++;
@@ -1730,7 +1732,7 @@ double precise_xstrtod(const char *str, char **endptr, char decimal, char sci,
     if (*p == decimal) {
         p++;
 
-        while (num_digits < max_digits && isdigit(*p)) {
+        while (num_digits < max_digits && isdigit_ascii(*p)) {
             number = number * 10. + (*p - '0');
             p++;
             num_digits++;
@@ -1738,7 +1740,7 @@ double precise_xstrtod(const char *str, char **endptr, char decimal, char sci,
         }
 
         if (num_digits >= max_digits)  // Consume extra decimal digits.
-            while (isdigit(*p)) ++p;
+            while (isdigit_ascii(*p)) ++p;
 
         exponent -= num_decimals;
     }
@@ -1752,7 +1754,7 @@ double precise_xstrtod(const char *str, char **endptr, char decimal, char sci,
     if (negative) number = -number;
 
     // Process an exponent string.
-    if (toupper(*p) == toupper(sci)) {
+    if (toupper_ascii(*p) == toupper_ascii(sci)) {
         // Handle optional sign
         negative = 0;
         switch (*++p) {
@@ -1765,7 +1767,7 @@ double precise_xstrtod(const char *str, char **endptr, char decimal, char sci,
         // Process string of digits.
         num_digits = 0;
         n = 0;
-        while (isdigit(*p)) {
+        while (isdigit_ascii(*p)) {
             n = n * 10 + (*p - '0');
             num_digits++;
             p++;
@@ -1798,7 +1800,7 @@ double precise_xstrtod(const char *str, char **endptr, char decimal, char sci,
 
     if (skip_trailing) {
         // Skip trailing whitespace.
-        while (isspace(*p)) p++;
+        while (isspace_ascii(*p)) p++;
     }
 
     if (endptr) *endptr = p;
@@ -1833,7 +1835,7 @@ int64_t str_to_int64(const char *p_item, int64_t int_min, int64_t int_max,
     int d;
 
     // Skip leading spaces.
-    while (isspace(*p)) {
+    while (isspace_ascii(*p)) {
         ++p;
     }
 
@@ -1846,7 +1848,7 @@ int64_t str_to_int64(const char *p_item, int64_t int_min, int64_t int_max,
     }
 
     // Check that there is a first digit.
-    if (!isdigit(*p)) {
+    if (!isdigit_ascii(*p)) {
         // Error...
         *error = ERROR_NO_DIGITS;
         return 0;
@@ -1865,7 +1867,7 @@ int64_t str_to_int64(const char *p_item, int64_t int_min, int64_t int_max,
                 if (d == tsep) {
                     d = *++p;
                     continue;
-                } else if (!isdigit(d)) {
+                } else if (!isdigit_ascii(d)) {
                     break;
                 }
                 if ((number > pre_min) ||
@@ -1878,7 +1880,7 @@ int64_t str_to_int64(const char *p_item, int64_t int_min, int64_t int_max,
                 }
             }
         } else {
-            while (isdigit(d)) {
+            while (isdigit_ascii(d)) {
                 if ((number > pre_min) ||
                     ((number == pre_min) && (d - '0' <= dig_pre_min))) {
                     number = number * 10 - (d - '0');
@@ -1902,7 +1904,7 @@ int64_t str_to_int64(const char *p_item, int64_t int_min, int64_t int_max,
                 if (d == tsep) {
                     d = *++p;
                     continue;
-                } else if (!isdigit(d)) {
+                } else if (!isdigit_ascii(d)) {
                     break;
                 }
                 if ((number < pre_max) ||
@@ -1916,7 +1918,7 @@ int64_t str_to_int64(const char *p_item, int64_t int_min, int64_t int_max,
                 }
             }
         } else {
-            while (isdigit(d)) {
+            while (isdigit_ascii(d)) {
                 if ((number < pre_max) ||
                     ((number == pre_max) && (d - '0' <= dig_pre_max))) {
                     number = number * 10 + (d - '0');
@@ -1931,7 +1933,7 @@ int64_t str_to_int64(const char *p_item, int64_t int_min, int64_t int_max,
     }
 
     // Skip trailing spaces.
-    while (isspace(*p)) {
+    while (isspace_ascii(*p)) {
         ++p;
     }
 
@@ -1954,7 +1956,7 @@ uint64_t str_to_uint64(uint_state *state, const char *p_item, int64_t int_max,
     int d;
 
     // Skip leading spaces.
-    while (isspace(*p)) {
+    while (isspace_ascii(*p)) {
         ++p;
     }
 
@@ -1968,7 +1970,7 @@ uint64_t str_to_uint64(uint_state *state, const char *p_item, int64_t int_max,
     }
 
     // Check that there is a first digit.
-    if (!isdigit(*p)) {
+    if (!isdigit_ascii(*p)) {
         // Error...
         *error = ERROR_NO_DIGITS;
         return 0;
@@ -1984,7 +1986,7 @@ uint64_t str_to_uint64(uint_state *state, const char *p_item, int64_t int_max,
             if (d == tsep) {
                 d = *++p;
                 continue;
-            } else if (!isdigit(d)) {
+            } else if (!isdigit_ascii(d)) {
                 break;
             }
             if ((number < pre_max) ||
@@ -1998,7 +2000,7 @@ uint64_t str_to_uint64(uint_state *state, const char *p_item, int64_t int_max,
             }
         }
     } else {
-        while (isdigit(d)) {
+        while (isdigit_ascii(d)) {
             if ((number < pre_max) ||
                 ((number == pre_max) && (d - '0' <= dig_pre_max))) {
                 number = number * 10 + (d - '0');
@@ -2012,7 +2014,7 @@ uint64_t str_to_uint64(uint_state *state, const char *p_item, int64_t int_max,
     }
 
     // Skip trailing spaces.
-    while (isspace(*p)) {
+    while (isspace_ascii(*p)) {
         ++p;
     }