@pep8speaks 2018-11-28
Hello @chris-b1! Thanks for submitting the PR.

 - There are no PEP8 issues in the file [`pandas/_version.py`](https://github.com/chris-b1/pandas/blob/6ef415532d180832219242ae623f7d4e1438a9d7/pandas/_version.py) !