@jreback 2018-11-28
can you add a whatsnew note in perf