@dragoljub 2018-11-28
@chris-b1 great work! 👍 The simple fix always seems to add a few more layers of complexity.

Can someone build/try this on Linux to make sure there is no regression there? 😉