@jbrockmendel 2018-12-02
Did tracking this problem down give any insight into whether/when we can de-duplicate parse_helper.xstrtod and tokenizer.xstrtod?