@chris-b1 2018-12-03
@jbrockmendel - It will take a bit of refactoring to deal with the `maybe_int` out param - the parse_helper one needs its, the csv parsing one doesn't.  Will have that as a follow-up, have #19361 on it.