@jreback 2018-12-05
looks like ci is fixed up, can you rebase